+++ pysmartnode/networking/wifi_esp8266.py
@@ -4,9 +4,10 @@ Created on 26.05.2018
 @author: Kevin Köck
 '''
 
-__updated__ = "2019-08-27"
+__updated__ = "2019-09-10"
 
 from pysmartnode import config
+from pysmartnode import logging
 import gc
 import uasyncio as asyncio
 import sys
@@ -20,6 +21,7 @@ if hasattr(config, "WIFI_SLEEP_MODE") and config.WIFI_SLEEP_MODE is not None:
 
 if hasattr(config, "RTC_SYNC_ACTIVE") and config.RTC_SYNC_ACTIVE is True:
     async def _sync():
+        s = 1
         while True:
             import ntptime
             print("Synchronize time from NTP server ...")
@@ -31,10 +33,15 @@ if hasattr(config, "RTC_SYNC_ACTIVE") and config.RTC_SYNC_ACTIVE is True:
                 machine.RTC().datetime(tm)
                 del ntptime
                 del sys.modules["ntptime"]
+                s = 1
                 await asyncio.sleep(18000)  # every 5h
             except Exception as e:
-                print("Error syncing time: {!s}, retry in 1s".format(e))
-                await asyncio.sleep(1)
+                await logging.getLogger("wifi").asyncLog("error",
+                                                         "Error syncing time: {!s}, retry in {!s}s".format(e, s))
+                await asyncio.sleep(s)
+                s += 5
+                # should prevent crashes because previous request was not finished and
+                # sockets still open (Errno 98 EADDRINUSE). Got killed by WDT after a few minutes.
 
 
     asyncio.get_event_loop().create_task(_sync())